refac folder structure
